+++ packages/queryable/index.ts
@@ -1,6 +1,5 @@
 import { jsS } from "@pnp/core";
 
-export * from "./invokable.js";
 export * from "./queryable.js";
 
 /**
@@ -15,10 +14,6 @@ export * from "./behaviors/inject-headers.js";
 export * from "./behaviors/parsers.js";
 export * from "./behaviors/timeout.js";
 export * from "./behaviors/resolvers.js";
-
-
-// testing
-export * from "./behaviors/diagnostics-jwt-decode.js";
 
 /**
  * Adds a property to a target instance
